servicios/integraciones/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
python-dotenv==1.0.1
pymongo==4.6.1
